feat(essensliste): moved useffect to component
